create helper function for JWT and import foruse in userController
